@@ -1,4 +1,3 @@
-﻿namespace BlazingChatter.Shared
-{
-    public record ActorAction(string User, bool IsTyping) : Actor(User);
-}
+﻿namespace BlazingChatter.Shared;
+
+public record ActorAction(string User, bool IsTyping) : Actor(User);
