@@ -1,9 +1,8 @@
 ﻿using BlazingChatter.Records;
 
-namespace BlazingChatter.Shared
-{
-    public record ActorCommand(
-        string User,
-        string OriginalText,
-        Command Command) : Actor(User);
-}
+namespace BlazingChatter.Shared;
+
+public record ActorCommand(
+    string User,
+    string OriginalText,
+    Command Command) : Actor(User);
