@@ -1,16 +1,15 @@
-﻿namespace BlazingChatter.Records
+﻿namespace BlazingChatter.Records;
+
+public record JokeApiResponse(string Type, Value Value);
+
+public class Value
 {
-    public record JokeApiResponse(string Type, Value Value);
+    public int Id { get; set; }
 
-    public class Value
+    public string? Joke
     {
-        public int Id { get; set; }
-
-        public string? Joke
-        {
-            get => _joke;
-            set => _joke = value?.Replace("&quot;", "\"");
-        }
-        private string? _joke;
+        get => _joke;
+        set => _joke = value?.Replace("&quot;", "\"");
     }
+    private string? _joke;
 }
