@@ -1,9 +1,9 @@
-﻿@inject NavigationManager Navigation
+﻿@using Microsoft.AspNetCore.Components.WebAssembly.Authentication;
+
+@inject NavigationManager Navigation
 
 @code {
-    protected override void OnInitialized()
-    {
-        Navigation.NavigateTo(
+    protected override void OnInitialized() => 
+        Navigation.NavigateToLogin(
             $"authentication/login?returnUrl={Uri.EscapeDataString(Navigation.Uri)}");
-    }
 }
