@@ -1,9 +1,6 @@
-﻿using System.Threading.Tasks;
+﻿namespace BlazingChatter.Services;
 
-namespace BlazingChatter.Services
+public interface ITranslationService
 {
-    public interface ITranslationService
-    {
-        ValueTask<(string text, bool isTranslated)> TranslateAsync(string text, string lang);
-    }
+    ValueTask<(string? text, bool isTranslated)> TranslateAsync(string text, string lang);
 }
